build: avoid narrowing conversions in initialization lists
Narrowing conversions in initialization lists are disallowed by the
standard, but nall was supressing this diagnostic (which is actually
an error by default). MSVC doesn't let you supress errors so easily, so
we'd do well to abide by the standard and perform explicit casts as
needed. For nall::shared_pointer::create() we can simply avoid the
uniform initialization syntax to bypass these extra rules, which is the
same strategy taken by implementations of std::make_shared().

+++ ares/ps1/gpu/gpu.hpp
@@ -207,9 +207,9 @@ struct GPU : Thread, Memory::Interface {
 
     static auto fromRGB(s32 r, s32 g, s32 b) -> Color {
       return {
-        std::clamp(r, 0, 255),
-        std::clamp(g, 0, 255),
-        std::clamp(b, 0, 255)
+        u8(std::clamp(r, 0, 255)),
+        u8(std::clamp(g, 0, 255)),
+        u8(std::clamp(b, 0, 255))
       };
     }
